Generate api new (#168)
* Remove API generation scripts.

* feat: Add Accept-language and timezone to headers

This change adds the `Accept-language` and `X-Timezone` headers to the HTTP client requests, utilizing the `LayoutService` for localization and timezone information.

* Refactor: Update tag management component to use signal and remove unnecessary imports.

* Regenerate api

* feat: Add date/time format and timezone to layout config

This change introduces configurable date and time formats, along with timezone support, to the application layout configuration. The configuration is loaded from local storage and defaults are provided.

* feat: Add date helper functions

This change introduces a set of helper functions for date manipulation and formatting, including string conversion, addition, and restoration from objects.

* Remove unused usings.

* feat: Add API generation settings

* Add API generation settings

This change adds `ApiGenerationSettings` to `AppSettings` in multiple projects.

* feat: Add API generation settings to appsettings

This change adds `ApiGenerationSettings` to `appsettings.json` and `appsettings.Development.json` in multiple projects.

* Refactoring

* Remove unused usings.

* Remove unused using.

* Add favicon.ico

* Fix: Use correct TagTypes enum

* Refactor: Replace BackgroundSyncService with KeycloakSyncBackgroundService

* Remove unused using statement.

* Refactor: Remove trailing whitespace.

* feat: Add startup tasks and runner

* feat: Generate TypeScript client from Swagger during startup

This task generates TypeScript clients from Swagger definitions during application startup if changes are detected. It uses `swagger-typescript-api` to generate the client code and saves it to the specified output path.

* feat: Add swagger generation and files

This commit adds swagger file generation and related files.

* feat: Add swagger web client generator documentation

Adds documentation for the swagger web client generator in both English and Russian.